+++ src/util/gitcommand.ts
@@ -1,7 +1,7 @@
 import { extensions } from "vscode";
 
 import { GIT_COMMAND_IN_PATH } from "../constants";
-import { VscodeGitExtension } from "../git.api.interface";
+import { VscodeGitExtension } from "../util/git.api.interface";
 
 export function getGitCommand(): string {
     const vscodeGit = extensions.getExtension<VscodeGitExtension>(